LayoutTests/fast/js/dfg-arguments-cross-code-origin-expected.txt

Tests uses of 'arguments' that are aliased but span code origins.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS baz(42) is 42
PASS successfullyParsed is true

TEST COMPLETE
